% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check package dependencies}
\usage{
checkDependencies(packageName = NULL, dependencyType = c("Imports", "Depends"))
}
\arguments{
\item{packageName}{Name of package to profile. If NULL current package}

\item{dependencyType}{Imports, depends, and/ or suggests}
}
\description{
Check package dependencies
}
