% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFileComplexity.R
\name{computeFileComplexity}
\alias{computeFileComplexity}
\title{computeFileComplexity}
\usage{
computeFileComplexity(filePath, fileExtension = "")
}
\arguments{
\item{filePath}{Path to the file to be investigated.}

\item{fileExtension}{Optional; Extension of the file. Supported types are:
1) R, 2) java, 3) cpp. If left blank, the file extension of the file defined
in the filePath parameter is used.}
}
\value{
Returns a data.frame with the following columns: 1) file, name of
the file investigated; 2) lines, number of lines of the investigated file;
3) complexity, complexity score.
}
\description{
Computes the complexity of a R, Java, or C++ file based on a point system.
The amount of points is the sum of the number of defined Functions;
Conditions used (if, else, while); iterators used (apply functions, for-
loops); switch-cases used; logical operators used (>, ==, <, etc.).
}
\examples{
filePath <- system.file(package = "DependencyReviewer", "testScript.R")
df <- computeFileComplexity(filePath)
}
