% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinLint.R
\name{darwinLintFile}
\alias{darwinLintFile}
\title{darwinLintFile}
\usage{
darwinLintFile(fileName)
}
\arguments{
\item{fileName}{Path to file to lint}
}
\description{
Lint a given file.
}
\examples{
\dontrun{
darwinLintFile("SomeFileName.R")
}
}
