% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseFunctionUse.R
\name{DEP_summariseFunctionUse}
\alias{DEP_summariseFunctionUse}
\title{DEP_summariseFunctionUse}
\usage{
DEP_summariseFunctionUse()
}
\value{
tibble
}
\description{
DEPRECATED Summarise functions used in R package
}
