% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyReviewer-package.R
\docType{package}
\name{DependencyReviewer-package}
\alias{DependencyReviewer}
\alias{DependencyReviewer-package}
\title{DependencyReviewer: Tool Suite to Investigate Other Packages}
\description{
Helps investigating other packages during code review by looking at their dependencies.
}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl}

}
\keyword{internal}
