% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFileComplexity.R
\name{getPatterns}
\alias{getPatterns}
\title{getPatterns}
\usage{
getPatterns(type)
}
\arguments{
\item{type}{File type, one of 'R', 'java', 'cpp'.}
}
\value{
Returns a list of patterns
}
\description{
Gets one of the supported patterns: 1) R, 2) java, 2) cpp.
}
