% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{makeGraph}
\alias{makeGraph}
\title{makeGraph}
\usage{
makeGraph(funsPerDefFun, width, height, pkgName, expFuns)
}
\arguments{
\item{funsPerDefFun}{}
}
\value{
diagram of the package
}
\description{
Makes the graph
}
