% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{runShiny}
\usage{
runShiny()
}
\value{
An object that represents the app.
}
\description{
Runs a Shiny app for dependency investigation.
}
\examples{
# Run only in interactive session
if (interactive()) {
  runShiny()
}
}
