% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseFunctionUse.R
\name{summariseFunctionUse}
\alias{summariseFunctionUse}
\title{Summarise functions used in R package}
\usage{
summariseFunctionUse()
}
\value{
tibble
}
\description{
Summarise functions used in R package
}
