% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinLint.R
\name{darwinLintPackage}
\alias{darwinLintPackage}
\title{darwinLintPackage}
\usage{
darwinLintPackage(path = ".")
}
\arguments{
\item{path}{Default: ".", Path to package}
}
\value{
List of lint objects.
}
\description{
Darwin lintr object, using default lintr object with camelCase
}
\examples{
darwinLintPackage()
}
