% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyReviewer-package.R
\docType{package}
\name{DependencyReviewer-package}
\alias{DependencyReviewer}
\alias{DependencyReviewer-package}
\title{DependencyReviewer: Tool suite to investigate other packages}
\description{
DependnecyReviewer is a package that helps investigating other packages during code review.
}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl}

}
\keyword{internal}
