% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyReviewer-package.R
\docType{package}
\name{DependencyReviewer-package}
\alias{DependencyReviewer}
\alias{DependencyReviewer-package}
\title{DependencyReviewer: Check package dependencise}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
