% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseFunctionUse.R
\name{funsUsedInFile}
\alias{funsUsedInFile}
\title{funsUsedInFile}
\usage{
funsUsedInFile(files)
}
\arguments{
\item{files}{Files to get functions from}
}
\value{
table
}
\description{
funsUsedInFile
}
