% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseFunctionUse.R
\name{funsUsedInFile}
\alias{funsUsedInFile}
\title{funsUsedInFile}
\usage{
funsUsedInFile(files, verbose = FALSE, in_package = TRUE)
}
\arguments{
\item{files}{Files to get functions from}

\item{verbose}{Verbosity}

\item{in_package}{default: TRUE}
}
\value{
table
}
\description{
Support function
}
