% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{pkgDiagram}
\alias{pkgDiagram}
\title{pkgDiagram}
\usage{
pkgDiagram(pkgPath, width = 1000, height = 1000, verbose = FALSE)
}
\value{
diagram image
}
\description{
Creates a diagram of all defined functions in a package.
}
\examples{
if (interactive()) {
  pkgDiagram(
    pkgPath = "./",
    width = 1000,
    height = 4000,
    verbose = TRUE)
}
}
